Move challenges into comicmodels
