+++ app/comicmodels/signals.py
@@ -4,7 +4,7 @@ from django.db.models.signals import post_save
 from django.dispatch import receiver
 from guardian.shortcuts import assign_perm
 
-from challenges.emails import send_challenge_created_email
+from comicmodels.emails import send_challenge_created_email
 from comicmodels.models import ComicSite
 from evaluation.models import Config
