+++ app/comicmodels/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import url
 
-from challenges.views import ChallengeCreate, ChallengeList, ChallengeUpdate
+from comicmodels.views import ChallengeCreate, ChallengeList, ChallengeUpdate
 
 urlpatterns = [
     url(r'^my-challenges/$', ChallengeList.as_view(), name='list'),
