+++ app/comicmodels/views.py
@@ -3,7 +3,7 @@ from django.contrib.messages.views import SuccessMessageMixin
 from django.db.models import Q
 from django.views.generic import CreateView, ListView, UpdateView
 
-from challenges.forms import ChallengeCreateForm, ChallengeUpdateForm
+from comicmodels.forms import ChallengeCreateForm, ChallengeUpdateForm
 from comicmodels.models import ComicSite
 from comicsite.permissions.mixins import UserIsChallengeAdminMixin
